Make brown actually respond to termination signals
Oh threads, thou art a cruel and confusing mistress.  Several things I
mucked up, including not setting the agent's class in the thread-local
variable until *after* the agent had run (which, of course, is never going
to work because the agent never quits), not signalling the agent runner
thread on stop, not terminating the agent thread join loop when there were
no agents, and, as a topper, restarting agents when they exit cleanly.